+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Added: Smarter interpretation of positional arguments passed to factory methods
+
 ## [0.1.0] - 2024-11-27
 
 - Initial release